feat: Validaciones de APIS y FrontEnd
- Se agregaron validaciones para Marca, Tipo de Producto y Producto para el correcto funcionamiento al momento de eliminar.

- Se agregó una validación para que al momento de crear un producto no admita valores negativos en el precio

+++ laravel/app/Http/Controllers/Api/ProductoController.php
@@ -74,6 +74,16 @@ class ProductoController extends Controller
             return response()->json(['message' => 'Producto no encontrado'], 404);
         }
 
+        if (!$producto['active']) {
+            return response()->json(['message' => 'error', 'message' => 'No se puede eliminar un producto que está activo'], 409);
+        }
+
+        $stock = intval($producto['stock']);
+
+        if ($stock > 0) {
+            return response()->json(['status' => 'error', 'message' => 'No se puede eliminar un producto con stock'], 409);
+        }
+
         $producto->delete();
 
         return response()->json(['message' => 'Producto eliminado']);
